moved custom collections to separate module
