+++ collection/build.gradle.kts
@@ -1,0 +1,8 @@
+dependencies {
+    implementation(project(":model"))
+    implementation(project(":logger"))
+}
+
+tasks.bootJar {
+    enabled = false
+}
