+++ collection/src/main/java/executor/service/collection/queue/QueueHandler.java
@@ -1,4 +1,4 @@
-package executor.service.publisher.queue;
+package executor.service.collection.queue;
 
 import java.util.List;
 import java.util.Optional;
