+++ settings.gradle.kts
@@ -1,3 +1,4 @@
 rootProject.name = "publisher_microservice"
 include("logger")
 include("model")
+include("collection")
