+++ src/main/java/executor/service/publisher/config/BeanConfiguration.java
@@ -1,25 +1,11 @@
 package executor.service.publisher.config;
 
-import executor.service.publisher.queue.ThreadSafeQueueHandler;
-import executor.service.publisher.queue.proxy.ProxySourceQueueHandler;
-import executor.service.publisher.queue.proxy.ProxySourceQueueHandlerImpl;
-import executor.service.publisher.queue.scenario.ScenarioSourceQueueHandler;
-import executor.service.publisher.queue.scenario.ScenarioSourceQueueHandlerImpl;
 import okhttp3.OkHttpClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class BeanConfiguration {
-    @Bean
-    public ProxySourceQueueHandler proxyQueueHandler() {
-        return new ProxySourceQueueHandlerImpl(new ThreadSafeQueueHandler<>());
-    }
-
-    @Bean
-    public ScenarioSourceQueueHandler scenarioQueueHandler() {
-        return new ScenarioSourceQueueHandlerImpl(new ThreadSafeQueueHandler<>());
-    }
 
     @Bean
     public OkHttpClient okHttpClient() {
